requirements.txt
# python-docx is a Python library for creating and updating Microsoft Word (.docx) files.
python-docx==1.1.2